TONS of changes to a bunch of systems

+++ src/core/player.cpp
@@ -91,7 +91,7 @@ void aiplayer_updateinput(struct aiplayer* p)
 
 	resetcontroller(p);
 
-	next_index = (p->vehicle->index_track + 3) % p->track->num_pathpoints;
+	next_index = (p->vehicle->index_track + 3) % (int)p->track->num_pathpoints;
 	vec3f_copy(next_point, p->track->pathpoints[next_index]);
 
 	// future point based on current speed
